+++ app/src/main/java/win/com/data/entity/ParticipantEntity.kt
@@ -5,8 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "participants")
 data class ParticipantEntity(
-    @PrimaryKey(autoGenerate = true) val id: Long = 0,
-    val eventId: Long,
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    val eventId: Int,
     val nickname: String,
     val team: String?,
     val role: String // HOST, PILOT, ORGA
